Update : alterado nome da TitulosMovimentacao para MovimentationType,criado alteado test para CreditController,alterado AccountMovementService verificar se o existe o customer

+++ Bootcamp TQI Kotlin - Backend Developer/api credit system/credit-aplication-sytem/src/main/resources/db/migration/V1__create_table_customer.sql
@@ -8,8 +8,8 @@ CREATE TABLE costumer (
    password VARCHAR(255) NOT NULL,
    zip_code VARCHAR(255) NOT NULL,
    street VARCHAR(255) NOT NULL,
-   account_free_balance DECIMAL,
-   balance_blocked DECIMAL,
+   account_free_balance FLOAT,
+   balance_blocked FLOAT,
    number_account VARCHAR(255),
    CONSTRAINT pk_costumer PRIMARY KEY (id)
 );
