+++ Bootcamp TQI Kotlin - Backend Developer/api credit system/credit-aplication-sytem/src/main/resources/db/migration/V2__create_table_credit.sql
@@ -1,7 +1,7 @@
 CREATE TABLE credit (
   id BIGINT AUTO_INCREMENT NOT NULL,
    credit_code char(36) NOT NULL,
-   credit_value DECIMAL NOT NULL,
+   credit_value FLOAT NOT NULL,
    day_first_installment BIGINT NOT NULL,
    number_of_installments INT NOT NULL,
    status INT NULL,
