+++ Bootcamp TQI Kotlin - Backend Developer/api credit system/credit-aplication-sytem/src/main/resources/db/migration/V3__table_acount_moviment.sql
@@ -3,7 +3,7 @@ CREATE TABLE account_movement (
    customer_id BIGINT,
    date_moviment BIGINT NOT NULL,
    type INT,
-   moviment_value DECIMAL,
+   moviment_value FLOAT,
    CONSTRAINT pk_account_movement PRIMARY KEY (id)
 );
